chore : js 파일 리팩토링 fix : mainPage.css 수정
- 함수들의 기능마다 나누어 분리
- cart 이미지 안보이는 현상 해결
- 할인 전 금액 안보이는 현상 해결
