+++ client/js/mainPage/animation/index.js
@@ -1,2 +1,5 @@
 export * from "./close.js"
 export * from "./swiper.js"
+export * from "./expand.js"
+export * from "./fade.js"
+export * from "./slide.js"
