+++ client/js/mainPage/index.js
@@ -1,1 +1,3 @@
-export * from '../mainPage/useData';
+export * from "./manageDate/index.js";
+export * from "./animation/index.js";
+export * from "./render/index.js";
