+++ client/js/mainPage/manageDate/index.js
@@ -1,0 +1,1 @@
+export * from "./useData.js";
