+++ client/js/mainPage/manageDate/useData.js
@@ -1,4 +1,4 @@
-import {tiger} from "../../lib/index.js"
+import {tiger} from "../../../lib/index.js"
 
 
 export const bannerData = (await tiger.get("http://localhost:3000/banner")).data;
