+++ client/js/mainPage/render/index.js
@@ -1,0 +1,1 @@
+export * from "./render.js";
